daily_video_app/requirements.txt
flask==3.0.2
huggingface_hub==0.20.3
google-auth==2.27.0
google-api-python-client==2.118.0
moviepy==1.0.3
python-dotenv==1.0.1
google-generativeai==0.3.2
requests==2.31.0
diffusers==0.25.0
torch==2.2.0
transformers==4.37.2
accelerate==0.27.2
gunicorn==21.2.0 